Fix bug in message sending
##Fix bug in message sending
This commit fixes a bug in the message sending process between the client and server. The bug caused some messages to be lost or not sent properly, leading to inconsistent communication between the two. This fix ensures that all messages are sent and received correctly, improving the overall stability and reliability of the chatbot communication system.